improved timeout handling

+++ src/utils/Poll.cpp
@@ -10,7 +10,7 @@
 namespace traceroute::utils
 {
 
-int Poll(std::vector<int> fds, int events, std::chrono::milliseconds timeout)
+int Poll(std::vector<int> fds, short int events, std::chrono::milliseconds timeout)
 {
     std::vector<pollfd> pollds(fds.size());
     for (int fd : fds)
